adding total management expenses unit test

+++ database/migrations/2023_09_03_102707_update_total_management_expenses_table_for_worker.php
@@ -12,9 +12,16 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('total_management_expenses', function (Blueprint $table) {
-            $table->dropColumn('quantity');
-            // Column for worker_id
-            $table->bigInteger('worker_id')->unsigned()->after('id')->change();
+            if (DB::getDriverName() !== 'sqlite') {
+                $table->dropColumn('quantity');
+                // Column for worker_id
+                $table->bigInteger('worker_id')->unsigned()->after('id')->change();
+            }
+            if (DB::getDriverName() === 'sqlite') {
+                $table->bigInteger('worker_id')->unsigned();
+                $table->string('type', 255)->nullable();
+                $table->decimal('deduction', 8,2)->default(0);
+            }
             // Column for application id
             $table->bigInteger('application_id')->unsigned();
             // Foreign key from total_management_applications table
@@ -23,12 +30,16 @@ return new class extends Migration
             $table->bigInteger('project_id')->unsigned();
             // Foreign key from total_management_project table
             $table->foreign('project_id')->references('id')->on('total_management_project')->onDelete('cascade');
-            // Column for Type
-            $table->string('type', 255)->nullable()->after('title')->change();
+            if (DB::getDriverName() !== 'sqlite') {
+                // Column for Type
+                $table->string('type', 255)->nullable()->after('title')->change();
+            }
             // Paid amount_paid column
             $table->decimal('amount_paid', 8,2)->default(0)->after('amount');
-            // Column for Deduction
-            $table->decimal('deduction', 8,2)->default(0)->after('amount_paid')->change();
+            if (DB::getDriverName() !== 'sqlite') {
+                // Column for Deduction
+                $table->decimal('deduction', 8,2)->default(0)->after('amount_paid')->change();
+            }
             // Column for remaining amount
             $table->decimal('remaining_amount', 8,2)->default(0)->after('deduction');
         });
